# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "react-ts backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/samples": {
            "get": {
                "tags": [
                    "samples"
                ],
                "summary": "実験用",
                "parameters": [
                    {
                        "type": "string",
                        "name": "q1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 3,
                        "type": "string",
                        "name": "q2",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "q3",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "q4",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "name": "q5",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "q6",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取得結果",
                        "schema": {
                            "$ref": "#/definitions/handler.GetSampleResponse"
                        }
                    },
                    "400": {
                        "description": "検証エラー",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/surveyors/query": {
            "get": {
                "tags": [
                    "surveyors"
                ],
                "summary": "指定した事業所の調査員のリストを返す",
                "parameters": [
                    {
                        "type": "string",
                        "description": "事業所のid",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "調査員のリスト",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.Surveyor"
                            }
                        }
                    },
                    "400": {
                        "description": "検証エラー",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "status bad request"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "handler.GetSampleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "000001"
                },
                "name": {
                    "type": "string",
                    "example": "調査員1"
                }
            }
        },
        "handler.Surveyor": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "000001"
                },
                "name": {
                    "type": "string",
                    "example": "調査員1"
                }
            }
        }
    }
}